+++ utils_a.c
@@ -6,7 +6,7 @@
 /*   By: csantos- <csantos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/07 21:11:12 by csantos-          #+#    #+#             */
-/*   Updated: 2021/04/14 21:52:51 by csantos-         ###   ########.fr       */
+/*   Updated: 2021/04/14 23:32:06 by csantos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ char		*ft_itoa(long int n)
 ** Converts input into hexadecimal x and X
 */
 
-char		*hextoa(t_flags *flags, unsigned long nb)
+char		*hextoa(t_flags *flags, unsigned long long nb)
 {
 	int	count;
 	char			*str;
